navigation and styling fixed
was missing some important styling for the ingredients pages
but that should be fixed now
Will have to fix the navbar now going all the way down the pages,
the current fix to do that messes with the ingredients side nav and causes that to be sticky for some reason
I'm not quite sure why it happens since the fixes to the styles aren't related to the side nav at all
Will fix some other time. I don't really thing that it's something that I need to fix right now.
It's an annoyance for sure, but not a huge deal.
I'll fix it when I have more time to look into it.
For now I want to merge the other two pages into this.
Hopefully the map won't break when I merge, but seeing how ingredients changed, it's possible.